tests/wpt/tests/html/canvas/element/canvas-host/2d.canvas.host.size.large.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.canvas.host.size.large</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.canvas.host.size.large</h1>
<p class="desc"></p>

<p class="notes">Not sure how reasonable this is, but the spec doesn't say there's an upper limit on the size.
<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

  var n = 2147483647; // 2^31 - 1, which should be supported by any sensible definition of "long"
  canvas.width = n;
  canvas.height = n;
  _assertSame(canvas.width, n, "canvas.width", "n");
  _assertSame(canvas.height, n, "canvas.height", "n");

});
</script>
